Enum iced_wgpu::Primitive

pub enum Primitive {
    None,
    Group {
        primitives: Vec<Primitive, Global>,
    },
    Text {
        content: String,
        bounds: Rectangle<f32>,
        color: Color,
        size: f32,
        font: Font,
        horizontal_alignment: Horizontal,
        vertical_alignment: Vertical,
    },
    Quad {
        bounds: Rectangle<f32>,
        background: Background,
        border_radius: f32,
        border_width: f32,
        border_color: Color,
    },
    Image {
        handle: Handle,
        bounds: Rectangle<f32>,
    },
    Svg {
        handle: Handle,
        bounds: Rectangle<f32>,
    },
    Clip {
        bounds: Rectangle<f32>,
        content: Box<Primitive, Global>,
    },
    Translate {
        translation: Vector<f32>,
        content: Box<Primitive, Global>,
    },
    Mesh2D {
        buffers: Mesh2D,
        size: Size<f32>,
    },
    Cached {
        cache: Arc<Primitive>,
    },
}

A rendering primitive.

Variants

None

An empty primitive

Group

Fields

primitives: Vec<Primitive, Global>

The primitives of the group

A group of primitives

Text

Fields

content: String

The contents of the text

bounds: Rectangle<f32>

The bounds of the text

color: Color

The color of the text

size: f32

The size of the text

font: Font

The font of the text

horizontal_alignment: Horizontal

The horizontal alignment of the text

vertical_alignment: Vertical

The vertical alignment of the text

A text primitive

Quad

Fields

bounds: Rectangle<f32>

The bounds of the quad

background: Background

The background of the quad

border_radius: f32

The border radius of the quad

border_width: f32

The border width of the quad

border_color: Color

The border color of the quad

A quad primitive

Image

Fields

handle: Handle

The handle of the image

bounds: Rectangle<f32>

The bounds of the image

An image primitive

Svg

Fields

handle: Handle

The path of the SVG file

bounds: Rectangle<f32>

The bounds of the viewport

An SVG primitive

Clip

Fields

bounds: Rectangle<f32>

The bounds of the clip

content: Box<Primitive, Global>

The content of the clip

A clip primitive

Translate

Fields

translation: Vector<f32>

The translation vector

content: Box<Primitive, Global>

The primitive to translate

A primitive that applies a translation

Mesh2D

Fields

buffers: Mesh2D

The vertex and index buffers of the mesh

size: Size<f32>

The size of the drawable region of the mesh.

Any geometry that falls out of this region will be clipped.

A low-level primitive to render a mesh of triangles.

It can be used to render many kinds of geometry freely.

Cached

Fields

cache: Arc<Primitive>

The cached primitive

A cached primitive.

This can be useful if you are implementing a widget where primitive generation is expensive.