iced_wgpu::window

Module compositor

Connect a window with a renderer.

Structs

Compositor

A window graphics backend for iced powered by wgpu.

Enums

Error

A compositor error.

Functions

new

Creates a Compositor with the given Settings and window.

present

Presents the given primitives with the given Compositor.

screenshot

Renders the current surface to an offscreen buffer.