Struct iced_wgpu::window::Compositor

pub struct Compositor<Theme> { /* private fields */ }

A window graphics backend for iced powered by wgpu.

Implementations

impl<Theme> Compositor<Theme>

pub async fn request<W: HasRawWindowHandle>(
settings: Settings,
compatible_window: Option<&W>
) -> Option<Self>

Requests a new Compositor with the given Settings.

Returns None if no compatible graphics adapter could be found.

pub fn create_backend(&self) -> Backend

Creates a new rendering Backend for this Compositor.